net: fec: the device id is tied to the enumeration order for mdio
So, the aliases allow the dev id to be consistent across device
tree overlays and fixups instead of relying on the order of the
nodes enumerating the platform devices. This can be seen when
the fec1 ends up as ethernet2 which is a problem because the
call to probe needs to ensure the mdio fec0_mii_bus is assigned
to the ethernet0.

This change uses the aliases if they are present in the device
tree to set the dev_id.

Signed-off-by: Charles Hardin <charles.hardin@chargepoint.com>